+++ src/misc/cache/_lv_cache_internal.h
@@ -76,28 +76,7 @@ struct _lv_cache_class_t {
 /**********************
  * GLOBAL PROTOTYPES
  **********************/
-void     lv_cache_entry_ref_reset(lv_cache_entry_t_ * entry);
-void     lv_cache_entry_ref_inc(lv_cache_entry_t_ * entry);
-void     lv_cache_entry_ref_dec(lv_cache_entry_t_ * entry);
-int32_t  lv_cache_entry_ref_get(lv_cache_entry_t_ * entry);
-uint32_t lv_cache_entry_get_generation(lv_cache_entry_t_ * entry);
-void     lv_cache_entry_inc_generation(lv_cache_entry_t_ * entry);
-void     lv_cache_entry_set_generation(lv_cache_entry_t_ * entry, uint32_t generation);
-uint32_t lv_cache_entry_get_node_size(lv_cache_entry_t_ * entry);
-void     lv_cache_entry_set_node_size(lv_cache_entry_t_ * entry, uint32_t node_size);
-void     lv_cache_entry_set_invalid(lv_cache_entry_t_ * entry, bool is_invalid);
-void     lv_cache_entry_set_cache(lv_cache_entry_t_ * entry, const lv_cache_t_ * cache);
-const lv_cache_t_ * lv_cache_entry_get_cache(const lv_cache_entry_t_ * entry);
-bool     lv_cache_entry_is_invalid(lv_cache_entry_t_ * entry);
-void  *  lv_cache_entry_get_data(lv_cache_entry_t_ * entry);
 
-void  *  lv_cache_entry_acquire_data(lv_cache_entry_t_ * entry);
-void     lv_cache_entry_release_data(lv_cache_entry_t_ * entry, void * user_data);
-lv_cache_entry_t_ * lv_cache_entry_get_entry(void * data, const uint32_t node_size);
-
-uint32_t lv_cache_entry_get_size(const uint32_t node_size);
-lv_cache_entry_t_ * lv_cache_entry_alloc(const uint32_t node_size);
-void lv_cache_entry_free(lv_cache_entry_t_ * entry);
 /*************************
  *    GLOBAL VARIABLES
  *************************/
