+++ src/misc/cache/_lv_cache_lru_rb.h
@@ -15,6 +15,7 @@ extern "C" {
  *      INCLUDES
  *********************/
 #include "_lv_cache_internal.h"
+#include "lv_cache_entry.h"
 
 /*********************
  *      DEFINES
