+++ src/misc/cache/lv_cache.h
@@ -15,6 +15,7 @@ extern "C" {
  *      INCLUDES
  *********************/
 #include "_lv_cache_internal.h"
+#include "lv_cache_entry.h"
 #include <stdbool.h>
 #include <stdlib.h>
